Fix CUDA fft template instantiation

+++ math/src/gpu/cuda/shaders/fft/fft.cuh
@@ -8,12 +8,12 @@ inline __device__ void _radix2_dit_butterfly(Fp *input,
 {
     if (blockIdx.x >= butterfly_count) return;
 
-    long group_count = 1 << stage;
-    long half_group_size = butterfly_count / group_count;
-    long group = threadIdx.x / half_group_size;
+    int group_count = 1 << stage;
+    int half_group_size = butterfly_count / group_count;
+    int group = threadIdx.x / half_group_size;
 
-    long pos_in_group = threadIdx.x % half_group_size;
-    long i = threadIdx.x * 2 - pos_in_group; // multiply quotient by 2
+    int pos_in_group = threadIdx.x % half_group_size;
+    int i = threadIdx.x * 2 - pos_in_group; // multiply quotient by 2
 
     Fp w = twiddles[group];
     Fp a = input[i];
